requirements.txt
python-telegram-bot==20.7
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openai==1.3.7
mistralai==0.1.3
schedule==1.2.0
# sqlite3 входит в стандартную библиотеку Python
python-dotenv==1.0.0
feedparser==6.0.10
Pillow==10.1.0
aiohttp==3.9.1
# asyncio входит в стандартную библиотеку Python
selenium==4.12.0
webdriver-manager==4.0.1
